requirements.txt
paho-mqtt
aiohttp
asyncio
certifi